Added some images and Created Navigator page
