+++ CampusNavigator/CampusNavigator/CampusNavigator/Components/Navigator/NaviagationSteps.swift
@@ -1,0 +1,23 @@
+import SwiftUI
+
+struct NavigationStepView: View {
+    let icon: String
+    let step: String
+    let time: String
+    
+    var body: some View {
+        HStack(alignment: .top) {
+            Image(systemName: icon)
+                .foregroundColor(.blue)
+                .frame(width: 30)
+            VStack(alignment: .leading) {
+                Text(step)
+                Text(time)
+                    .font(.caption)
+                    .foregroundColor(.gray)
+            }
+            Spacer()
+        }
+        .padding(.vertical, 4)
+    }
+}
